#818 added usage stats probe and consolidated shared code into base probe
